requirements.txt
Django>=5.0
django-unfold>=0.30.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-celery-beat>=2.5.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
django-ses>=3.5.0
pycryptodome>=3.20.0
requests>=2.31.0
gunicorn>=21.2.0
